chore: last refactoring
