+++ packages/shared/src/components/auth/OnboardingSteps/index.ts
@@ -1,3 +1,3 @@
 export * from './EditTag';
 export * from './ReadingReminder';
-export * from './ContentTypes';
+export * from './ContentTypes/ContentTypes';
